engine: three latent UB defects surfaced by making UBSan non-recoverable; fix all three + land the control
The asan presets have carried -fsanitize=address,undefined since the
Windows-preset work, but UBSan's DEFAULT is print-and-continue and ctest
never fails on a recoverable diagnostic -- so "asan preset green" never
actually asserted UB-freedom. Turning on -fno-sanitize-recover=undefined
(cmake/Sanitizers.cmake, both toolchain branches) surfaced three real
engine defects, each previously invisible because the tests that reached
them either mirrored the bug or never checked the poisoned output:

1. event_flag_set/event_flag_test (event_framework.cpp:257/:268) shifted
   the hi FIRED word by (id - 33u), which UNDERFLOWS at id 32 -- shift
   count 0xFFFFFFFF, formally UB. x86's shift masking parked id 32 on hi
   bit 31, and the round-trip test + translator pin mirrored the same
   expression, so everything stayed green over UB. The layout is now the
   intended (id - 32): hi word bit 0 is id 32, tiling all 32 bits.
   Nothing persisted carries a nonzero hi word (Act-1 captures cannot
   fire ids >= 32; the twin fixtures and 50-seed corpus are Act-1), so no
   golden moved -- the suite proves it. Comments repinned at
   event_framework.hpp, run_state.hpp, public_view.hpp, translate.cpp,
   seed_scan_test.cpp; test pins at event_framework_test.cpp (round trip
   now deliberately does NOT mirror the accessor's expression),
   translator_test.cpp (Beggar id 34 -> bit 2).

2. floor_stream/map_stream (rng_stream.hpp:247/:260) computed
   `run_seed + offset` in signed int64 -- UB exactly where golden set 5
   pins the edge: a Long.MAX_VALUE seed. Java's long addition wraps, and
   the golden bytes were matching only via the accidental wrap. Both
   helpers (and the test's own open-coded cross-check at
   rng_stream_test.cpp:312) now wrap THROUGH uint64_t -- defined, and
   bit-identical to Java. No behavior change at any seed.

3. The MAP_CHOICE legal-action mask arm (run_advance.cpp:2266) indexed
   rs.map[run_state_map_index(x, y)] with only the low guard: run_cur_row
   saturates at -1 below but not above, so an off-nominal (act, floor)
   pair -- act 0's base is -17, hence a value-init controller
   phase-flipped to MAP_CHOICE reads row 16 = map[112], out of bounds by
   7 -- fed garbage edges into the mask. The arm now takes the same
   `y >= kMapRows` high guard map_edge_connects has always had (offer
   nothing), because the mask is a total pull API over arbitrary
   controllers; the transition path keeps its assert. In every reachable
   state MAP_CHOICE has row <= kMapRows-2, so real runs are unaffected.

All six presets green after the change -- debug/asan/release (WSL GCC)
and win-debug/win-asan/win-release (clang-cl), 2600 tests each, with the
asan pair now running UBSan in non-recoverable mode on both toolchains.
check_stale_counts + check_doc_links clean.

+++ include/sts/engine/rng_stream.hpp
@@ -244,7 +244,14 @@ constexpr void advance_counter_to(RngStream& s, int32_t target) noexcept {
 // at floor-transition time. Counter starts at 0.
 [[nodiscard]] constexpr RngStream floor_stream(int64_t run_seed,
                                                int32_t floor_after_increment) noexcept {
-    return from_seed(run_seed + static_cast<int64_t>(floor_after_increment));
+    // The sum wraps THROUGH uint64_t: Java's `seed + floorNum` is a long
+    // addition with defined two's-complement wraparound, and golden set 5
+    // pins a run_seed of Long.MAX_VALUE where `+ floor` overflows. Signed
+    // overflow in C++ is UB (UBSan flagged exactly this line); the unsigned
+    // round trip is the bit-identical, defined spelling of Java's wrap.
+    return from_seed(static_cast<int64_t>(
+        static_cast<uint64_t>(run_seed) +
+        static_cast<uint64_t>(static_cast<int64_t>(floor_after_increment))));
 }
 
 // Act-scoped mapRng seed (design doc §3.4; Exordium/TheCity/TheBeyond/
@@ -257,7 +264,10 @@ constexpr void advance_counter_to(RngStream& s, int32_t target) noexcept {
     constexpr int64_t kActMult[5] = {0, 1, 100, 200, 300};
     const int64_t offset =
         static_cast<int64_t>(act_num) * kActMult[act_num];
-    return from_seed(run_seed + offset);
+    // Same defined-wrap spelling as floor_stream above: Java's long addition
+    // wraps, C++'s signed addition must not be allowed to.
+    return from_seed(static_cast<int64_t>(static_cast<uint64_t>(run_seed) +
+                                          static_cast<uint64_t>(offset)));
 }
 
 }  // namespace sts::engine
